Separating online application search into anothe class
Online application search is not using pg_search and is returning only single result. I decide to separate it into different class so it's clear what it is about.
