+++ app/services/application_search.rb
@@ -2,20 +2,15 @@ class ApplicationSearch
   include Rails.application.routes.url_helpers
   attr_reader :error_message, :results
 
+  SORT_PARAMS = { reference: 'applications.reference', entered: 'applications.created_at',
+                  first_name: 'applicants.first_name', last_name: 'applicants.last_name',
+                  case_number: 'details.case_number', fee: 'details.fee',
+                  remission: 'applications.decision_cost',
+                  completed: 'applications.decision_date' }.freeze
+
   def initialize(query, current_user)
     @query = query
     @current_user = current_user
-  end
-
-  def online
-    return if !prepare_reference! || application_exists_and_user_can_access ||
-              application_exists_and_user_cannot_access || online_application_income_invalid?
-
-    if online_application_exists
-      edit_online_application_path(@online_application)
-    else
-      set_error_and_return_nil(:not_found)
-    end
   end
 
   def completed
@@ -44,68 +39,8 @@ class ApplicationSearch
       given_office_only(@current_user.office_id)
   end
 
-  def prepare_reference!
-    if @query.present?
-      reference = @query.upcase
-      reference.gsub!('HWF', '')
-      reference.gsub!(/[- ]/, '')
-      @query = "HWF-#{reference.scan(/.{1,3}/).join('-')}"
-    end
-  end
-
-  def application_exists_and_user_can_access
-    if application_exists && user_can_access
-      redirect_data = CompletedApplicationRedirect.new(@application)
-      @error_message = I18n.t(:processed_html, scope: scope, application_path: redirect_data.path)
-    end
-  end
-
-  def application_exists_and_user_cannot_access
-    if application_exists && !user_can_access
-      @error_message = I18n.t(:processed_by, scope: scope, office_name: application_office)
-    end
-  end
-
-  def application_exists
-    @application ||= Application.find_by(reference: @query.upcase)
-  end
-
-  def user_can_access
-    Pundit.policy(@current_user, @application).show?
-  end
-
-  def online_application_exists
-    @online_application ||= OnlineApplication.find_by(reference: @query.upcase)
-  end
-
-  def online_application_income_invalid?
-    if online_application_exists && income_required_but_missing
-      @error_message = I18n.t(:income_error, scope: scope)
-    end
-  end
-
-  def income_required_but_missing
-    online_application_income_required? && online_application_income_missing?
-  end
-
-  def online_application_income_required?
-    @online_application.benefits.eql?(false)
-  end
-
-  def online_application_income_missing?
-    [
-      @online_application.income,
-      @online_application.income_min_threshold_exceeded,
-      @online_application.income_max_threshold_exceeded
-    ].all?(&:nil?)
-  end
-
   def scope
     'activemodel.errors.models.forms/search.attributes.reference'
-  end
-
-  def application_office
-    @application.office.name
   end
 
   def set_error_and_return_nil(i18n_key, i18n_params = {})
@@ -134,23 +69,6 @@ class ApplicationSearch
   end
 
   def new_sort_param
-    case @sort_by
-    when 'reference'
-      "applications.reference #{@sort_to}"
-    when 'entered'
-      "applications.created_at #{@sort_to}"
-    when 'first_name'
-      "applicants.first_name #{@sort_to}"
-    when 'last_name'
-      "applicants.last_name #{@sort_to}"
-    when 'case_number'
-      "details.case_number #{@sort_to}"
-    when 'fee'
-      "details.fee #{@sort_to}"
-    when 'remission'
-      "applications.decision_cost #{@sort_to}"
-    when 'completed'
-      "applications.decision_date #{@sort_to}"
-    end
+    SORT_PARAMS[@sort_by.to_sym] + " #{@sort_to}"
   end
 end
